Implement review suggestions for openapi spec Diet now array & username can have dots

+++ services/py-help-service/client/cooking_assistant_api_client/models/recipe.py
@@ -22,7 +22,7 @@ class Recipe:
         title (str):
         ingredients (list[RecipeIngredient]):
         instructions (list[str]):
-        portions (int):
+        portions (float):
         id (int):
         nutrients (RecipeNutrients | Unset):
     """
@@ -30,7 +30,7 @@ class Recipe:
     title: str
     ingredients: list[RecipeIngredient]
     instructions: list[str]
-    portions: int
+    portions: float
     id: int
     nutrients: RecipeNutrients | Unset = UNSET
